+++ CHANGELOG.md
@@ -35,6 +35,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - add subnet resource config *subnet.tf*
 - add dhcp resource config *dhcp.tf*
 - add provider resource config *provider.tf*
+- add home region config *home_region.tf*
 
 ### Changed